refactor(export): remove LBExportFile and move suggestedFileName to LBExportFormat
